+++ iogt/settings/base.py
@@ -26,6 +26,7 @@ BASE_DIR = os.path.dirname(PROJECT_DIR)
 INSTALLED_APPS = [
     'home',
     'search',
+    'polls',
 
     'wagtail.contrib.forms',
     'wagtail.contrib.redirects',
